Added basic test infrastructure
