+++ Sources/vapor-main/main.swift
@@ -7,28 +7,7 @@
 #endif
 
 import Foundation
-
-let version = "0.6.0"
-
-struct VaporCLI {
-    // this closure assignment is necessary to be able to exclude Xcode on Linux
-    static let commands: [Command.Type] = {
-        var c = [Command.Type]()
-        c.append(Help)
-        c.append(Version)
-        c.append(Clean)
-        c.append(Build)
-        c.append(Run)
-        c.append(New)
-        c.append(Update)
-        #if os(OSX)
-            c.append(Xcode)
-        #endif
-        c.append(Heroku)
-        c.append(Docker)
-        return c
-    }()
-}
+import vapor
 
 var iterator = Process.arguments.makeIterator()
 
@@ -38,7 +17,7 @@ guard let directory = iterator.next() else {
     fail("no directory")
 }
 guard let commandId = iterator.next() else {
-    print("Usage: \(directory) [\(VaporCLI.commands.map({ $0.id }).joined(separator: "|"))]")
+    // print("Usage: \(directory) [\(VaporCLI.commands.map({ $0.id }).joined(separator: "|"))]")
     fail("no command")
 }
 guard let command = getCommand(id: commandId, commands: VaporCLI.commands) else {
