+++ Tests/LinuxMain.swift
@@ -1,0 +1,6 @@
+import XCTest
+@testable import vaporTestSuite
+
+XCTMain([
+     testCase(ArrayExtTests.allTests),
+])
