+++ test
@@ -1,0 +1,27 @@
+#!/bin/bash
+
+# See http://finestructure.co/blog/2016/6/20/serverless-continuous-integration
+# for the motivation for this script
+
+LOGFILE=test.log
+set -euo pipefail
+
+function build_and_test() {
+    date 2>&1 | tee -a $LOGFILE
+    uname -a 2>&1 | tee -a $LOGFILE
+    swift build --clean && swift build && swift test 2>&1 | tee -a $LOGFILE
+}
+
+if [ ! -f /.dockerenv ]; then
+    # We're not in a docker container
+    # this is where we start when run from the host
+    rm $LOGFILE
+    build_and_test
+    # we're going to call this same script from inside a container next
+    # to run the Linux tests, effectively jumping to the else branch
+    IMG=qutheory/swift:$(cat .swift-version)
+    docker run --rm -it -v $PWD:/vapor --entrypoint /vapor/test $IMG
+else
+    # Now we're in a docker container and run the Linux tests
+    build_and_test
+fi
